moving unit test to internal pkg
