+++ internal/otel/otel_test.go
@@ -264,3 +264,58 @@ func TestContextWithSpan(t *testing.T) {
 		t.Errorf("got wrong span")
 	}
 }
+
+func TestAddTraceID(t *testing.T) {
+	tests := []struct {
+		name       string
+		enabled    bool
+		hasTraceID bool
+		wantHeader bool
+	}{
+		{
+			name:       "otel enabled with trace id",
+			enabled:    true,
+			hasTraceID: true,
+			wantHeader: true,
+		},
+		{
+			name:       "otel enabled without trace id",
+			enabled:    true,
+			hasTraceID: false,
+			wantHeader: false,
+		},
+		{
+			name:       "otel disabled",
+			enabled:    false,
+			hasTraceID: false,
+			wantHeader: false,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			req := httptest.NewRequest(http.MethodGet, "/", nil)
+			w := httptest.NewRecorder()
+
+			otelConfig := Config{
+				Enabled:  tt.enabled,
+				Exporter: "http",
+				Endpoint: "http://localhost:4317",
+			}
+
+			if tt.hasTraceID {
+				ot := InitOpenTelemetry(context.Background(), logrus.New(), &otelConfig, "test", "test", false, "test", false, []string{})
+				ctx, _ := ot.Tracer().Start(context.Background(), "testing")
+				req = req.WithContext(ctx)
+			}
+
+			AddTraceID(w, req, tt.enabled)
+
+			if tt.wantHeader && w.Header().Get("X-Tyk-Trace-Id") == "" {
+				t.Errorf("expected header to be set, but it wasn't")
+			} else if !tt.wantHeader && w.Header().Get("X-Tyk-Trace-Id") != "" {
+				t.Errorf("expected header not to be set, but it was")
+			}
+		})
+	}
+}
